Extract transaction rows
